sort SeafRepos inside of SeafGroup
